Crate ena

An implementation of union-find. See the unify module for more details.

Modules

snapshot_vec

A utility class for implementing “snapshottable” things; a snapshottable data structure permits you to take a snapshot (via start_snapshot) and then, after making some changes, elect either to rollback to the start of the snapshot or commit those changes.

undo_log

Module which contains the snapshot/rollback functionality of the ena data structures.

unify

Union-find implementation. The main type is UnificationTable.