Module ena::config

Configuration.

Used to parse the config file or read from environment variables.
Also supplied are various helper functions for a CLI program.

Structs

BoardSettings
Settings for a board

Config
Upper level configuration hold json fields and values

ConfigInner
Use a deserialized BoardSettings as a base for all other BoardSettings Another struct is used to prevent a recursive loop in deserialization

Settings
Archiver settings

Statics

CONFIG_CONTENTS

Functions

check_version
Check to see if the first argument passed to the program is -v or --version

default_headers
Create the default headers for reqwest::Client

display
Display an ascii art with the crate version

display_full_version
Display the build information

display_help

ena_resume
Return the value of the ENA_RESUME environment variable

read_config
Read a Config file

read_json
Read a json file

read_json_try
Safe read a json file

refresh_rate
Create an iterator that mimics the thread refresh system

version
Return the current version of the crate