Struct ena::snapshot_vec::SnapshotVec

pub struct SnapshotVec<D: SnapshotVecDelegate> { /* fields omitted */ }

Methods

impl<D: SnapshotVecDelegate> SnapshotVec<D>

pub fn new() -> Self

pub fn with_capacity(c: usize) -> SnapshotVec<D>

pub fn record(&mut self, action: D::Undo)

pub fn len(&self) -> usize

pub fn push(&mut self, elem: D::Value) -> usize

pub fn get(&self, index: usize) -> &D::Value

pub fn reserve(&mut self, additional: usize)

Reserve space for new values, just like an ordinary vec.

pub fn get_mut(&mut self, index: usize) -> &mut D::Value

Returns a mutable pointer into the vec; whatever changes you make here cannot be undone automatically, so you should be sure call record() with some sort of suitable undo action.

pub fn set(&mut self, index: usize, new_elem: D::Value)

Updates the element at the given index. The old value will saved (and perhaps restored) if a snapshot is active.

pub fn set_all(&mut self, new_elems: impl FnMut(usize) -> D::Value)

Updates all elements. Potentially more efficient -- but otherwise equivalent to -- invoking set for each element.

pub fn update<OP>(&mut self, index: usize, op: OP) where
OP: FnOnce(&mut D::Value),
D::Value: Clone,

pub fn start_snapshot(&mut self) -> Snapshot

pub fn actions_since_snapshot(&self, snapshot: &Snapshot) -> &[UndoLog<D>]

pub fn rollback_to(&mut self, snapshot: Snapshot)

pub fn commit(&mut self, snapshot: Snapshot)

Commits all changes since the last snapshot. Of course, they can still be undone if there is a snapshot further out.