Trait ena::sql::Archiver

pub trait Archiver: Sync + Send {
#[must_use]
    fn run_inner<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Implementation details of archiving

The logic/algorithm in how you want to approach downloading everything.
With that said, it should have more methods and stuff but I just use it as
a way to pass generic implementations of YotsubaArchiver for different databases.

Required methods

#[must_use] fn run_inner<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Loading content...