Trait ena::undo_log::Snapshots

pub trait Snapshots<T>: UndoLogs<T> {
    type Snapshot;

    fn actions_since_snapshot(&self, snapshot: &Self::Snapshot) -> &[T];
    fn start_snapshot(&mut self) -> Self::Snapshot;
    fn rollback_to<R>(
        &mut self, 
        storage: impl FnOnce() -> R, 
        snapshot: Self::Snapshot
    )
    where
        R: Rollback<T>;
    fn commit(&mut self, snapshot: Self::Snapshot);

    fn has_changes(&self, snapshot: &Self::Snapshot) -> bool { ... }
}

A trait which extends UndoLogs to allow snapshots to be done at specific points. Each snapshot can then be used to rollback any changes to an underlying data structures if they were not desirable.

Each snapshot must be consumed linearly with either rollback_to or commit.

Required Associated Types

type Snapshot

Required Methods

fn actions_since_snapshot(&self, snapshot: &Self::Snapshot) -> &[T]

Returns the slice of actions that were taken since the snapshot began.

fn start_snapshot(&mut self) -> Self::Snapshot

Starts a new snapshot. That snapshot must eventually either be committed via a call to commit or rollback via rollback_to. Snapshots can be nested (i.e., you can start a snapshot whilst another snapshot is in progress) but you must then commit or rollback the inner snapshot before attempting to commit or rollback the outer snapshot.

fn rollback_to<R>(
&mut self,
storage: impl FnOnce() -> R,
snapshot: Self::Snapshot
) where
R: Rollback<T>,

Rollback (undo) the changes made to storage since the snapshot.

fn commit(&mut self, snapshot: Self::Snapshot)

Commit: keep the changes that have been made since the snapshot began

Provided Methods

fn has_changes(&self, snapshot: &Self::Snapshot) -> bool

Returns true if self has made any changes since snapshot started.