Struct ena::unify::UnificationTable

pub struct UnificationTable<S: UnificationStore> { /* fields omitted */ }

Table of unification keys and their values. You must define a key type K that implements the UnifyKey trait. Unification tables can be used in two-modes:

in-place (UnificationTable<InPlace<K>> or InPlaceUnificationTable<K>):

This is the standard mutable mode, where the array is modified in place.

To do backtracking, you can employ the snapshot and rollback_to methods.

persistent (UnificationTable<Persistent<K>> or PersistentUnificationTable<K>):

In this mode, we use a persistent vector to store the data, so that cloning the table is an O(1) operation.

This implies that ordinary operations are quite a bit slower though.

Requires the persistent feature be selected in your Cargo.toml file.

Methods

impl<S: UnificationStore> UnificationTable<S>

pub fn new() -> Self

pub fn snapshot(&mut self) -> Snapshot<S>

Starts a new snapshot. Each snapshot must be either rolled back or committed in a "LIFO" (stack) order.

pub fn rollback_to(&mut self, snapshot: Snapshot<S>)

Reverses all changes since the last snapshot. Also removes any keys that have been created since then.

pub fn commit(&mut self, snapshot: Snapshot<S>)

Commits all changes since the last snapshot. Of course, they can still be undone if there is a snapshot further out.

pub fn new_key(&mut self, value: S::Value) -> S::Key

Creates a fresh key with the given value.

pub fn reserve(&mut self, num_new_keys: usize)

Reserve memory for num_new_keys to be created. Does not actually create the new keys; you must then invoke new_key.

pub fn reset_unifications(
&mut self,
value: impl FnMut(S::Key) -> S::Value
)

Clears all unifications that have been performed, resetting to the initial state. The values of each variable are given by the closure.

pub fn len(&self) -> usize

Returns the number of keys created so far.

impl<'tcx, S, K, V> UnificationTable<S> where
S: UnificationStore<Key = K, Value = V>,
K: UnifyKey<Value = V>,
V: UnifyValue,

//////////////////////////////////////////////////////////////////////// Public API

pub fn union<K1, K2>(&mut self, a_id: K1, b_id: K2) where
K1: Into<K>,
K2: Into<K>,
V: UnifyValue<Error = NoError>,

Unions two keys without the possibility of failure; only applicable when unify values use NoError as their error type.

pub fn union_value<K1>(&mut self, id: K1, value: V) where
K1: Into<K>,
V: UnifyValue<Error = NoError>,

Unions a key and a value without the possibility of failure; only applicable when unify values use NoError as their error type.

pub fn unioned<K1, K2>(&mut self, a_id: K1, b_id: K2) -> bool where
K1: Into<K>,
K2: Into<K>,

Given two keys, indicates whether they have been unioned together.

pub fn find<K1>(&mut self, id: K1) -> K where
K1: Into<K>,

Given a key, returns the (current) root key.

pub fn unify_var_var<K1, K2>(
&mut self,
a_id: K1,
b_id: K2
) -> Result<(), V::Error> where
K1: Into<K>,
K2: Into<K>,

Unions together two variables, merging their values. If merging the values fails, the error is propagated and this method has no effect.

pub fn unify_var_value<K1>(&mut self, a_id: K1, b: V) -> Result<(), V::Error> where
K1: Into<K>,

Sets the value of the key a_id to b, attempting to merge with the previous value.

pub fn probe_value<K1>(&mut self, id: K1) -> V where
K1: Into<K>,

Returns the current value for the given key. If the key has been union'd, this will give the value from the current root.